Create PIN Finding 1
Suppose we  are provided with 4 numbers input1, input2, input and imput4 and are expected to find a secret PIN or Key.The three numbers input1, input2 and input3 are four digit numbers within the range 1000 and <- 9999. input4 is a positive integer number.
Calcilate the PIN using the below formula:
PIN= (largest digit in input1) x (largest digit in input2) x (largest digit in input3)) + input4
Here is a Java code for the above mentioned problem.